# src/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "CAPTION CUỐI KHÓA"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/room": {
      "post": {
        "tags": [
          "Room"
        ],
        "operationId": "createNewRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createNewRoomRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "tags": [
          "Room"
        ],
        "operationId": "getAllRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/room/{id}": {
      "get": {
        "tags": [
          "Room"
        ],
        "operationId": "getRoomById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "tags": [
          "Room"
        ],
        "operationId": "updateRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateRoomRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "Room"
        ],
        "operationId": "deleteRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/room/uploads-image": {
      "post": {
        "tags": [
          "Room"
        ],
        "operationId": "uploadImageRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "file",
            "format": "file",
            "description": ""
          },
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/api/location": {
      "get": {
        "tags": [
          "Location"
        ],
        "operationId": "getAllLocation",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "tags": [
          "Location"
        ],
        "operationId": "addNewLocation",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/addNewLocationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/location/get-location-by-page": {
      "get": {
        "tags": [
          "Location"
        ],
        "operationId": "getLocationByPage",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageIndex",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/location/{id}": {
      "get": {
        "tags": [
          "Location"
        ],
        "operationId": "getViTriById",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "tags": [
          "Location"
        ],
        "operationId": "deleteViTri",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/api/location/uploads-image": {
      "post": {
        "tags": [
          "Location"
        ],
        "operationId": "uploadImageLocation",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": true,
            "type": "file",
            "format": "file",
            "description": ""
          },
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createNewRoomRequest": {
      "title": "createNewRoomRequest",
      "example": {
        "ten_phong": "Phòng Deluxe",
        "phong_khach": 1,
        "phong_ngu": 1,
        "giuong": 2,
        "phong_tam": 1,
        "mo_ta": "Phòng sang trọng với đầy đủ tiện nghi",
        "gia_tien": 150,
        "may_giat": true,
        "ban_la": true,
        "tivi": true,
        "dieu_hoa": true,
        "wifi": true,
        "bep": true,
        "do_xe": true,
        "ho_boi": true,
        "ban_ui": true,
        "hinh_anh": "https://example.com/room.jpg",
        "vi_tri_id": 2
      },
      "type": "object",
      "properties": {
        "ten_phong": {
          "type": "string"
        },
        "phong_khach": {
          "type": "integer",
          "format": "int32"
        },
        "phong_ngu": {
          "type": "integer",
          "format": "int32"
        },
        "giuong": {
          "type": "integer",
          "format": "int32"
        },
        "phong_tam": {
          "type": "integer",
          "format": "int32"
        },
        "mo_ta": {
          "type": "string"
        },
        "gia_tien": {
          "type": "integer",
          "format": "int32"
        },
        "may_giat": {
          "type": "boolean"
        },
        "ban_la": {
          "type": "boolean"
        },
        "tivi": {
          "type": "boolean"
        },
        "dieu_hoa": {
          "type": "boolean"
        },
        "wifi": {
          "type": "boolean"
        },
        "bep": {
          "type": "boolean"
        },
        "do_xe": {
          "type": "boolean"
        },
        "ho_boi": {
          "type": "boolean"
        },
        "ban_ui": {
          "type": "boolean"
        },
        "hinh_anh": {
          "type": "string"
        },
        "vi_tri_id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ten_phong",
        "phong_khach",
        "phong_ngu",
        "giuong",
        "phong_tam",
        "mo_ta",
        "gia_tien",
        "may_giat",
        "ban_la",
        "tivi",
        "dieu_hoa",
        "wifi",
        "bep",
        "do_xe",
        "ho_boi",
        "ban_ui",
        "hinh_anh",
        "vi_tri_id"
      ]
    },
    "updateRoomRequest": {
      "title": "updateRoomRequest",
      "example": {
        "ten_phong": "Phòng Deluxe",
        "phong_khach": 1,
        "phong_ngu": 1,
        "giuong": 2,
        "phong_tam": 2,
        "mo_ta": "Phòng sang trọng với đầy đủ tiện nghi",
        "gia_tien": 1500000,
        "may_giat": true,
        "ban_la": true,
        "tivi": true,
        "dieu_hoa": true,
        "wifi": true,
        "bep": true,
        "do_xe": true,
        "ho_boi": true,
        "ban_ui": true,
        "vi_tri_id": 1
      },
      "type": "object",
      "properties": {
        "ten_phong": {
          "type": "string"
        },
        "phong_khach": {
          "type": "integer",
          "format": "int32"
        },
        "phong_ngu": {
          "type": "integer",
          "format": "int32"
        },
        "giuong": {
          "type": "integer",
          "format": "int32"
        },
        "phong_tam": {
          "type": "integer",
          "format": "int32"
        },
        "mo_ta": {
          "type": "string"
        },
        "gia_tien": {
          "type": "integer",
          "format": "int32"
        },
        "may_giat": {
          "type": "boolean"
        },
        "ban_la": {
          "type": "boolean"
        },
        "tivi": {
          "type": "boolean"
        },
        "dieu_hoa": {
          "type": "boolean"
        },
        "wifi": {
          "type": "boolean"
        },
        "bep": {
          "type": "boolean"
        },
        "do_xe": {
          "type": "boolean"
        },
        "ho_boi": {
          "type": "boolean"
        },
        "ban_ui": {
          "type": "boolean"
        },
        "vi_tri_id": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "ten_phong",
        "phong_khach",
        "phong_ngu",
        "giuong",
        "phong_tam",
        "mo_ta",
        "gia_tien",
        "may_giat",
        "ban_la",
        "tivi",
        "dieu_hoa",
        "wifi",
        "bep",
        "do_xe",
        "ho_boi",
        "ban_ui",
        "vi_tri_id"
      ]
    },
...
    "addNewLocationRequest": {
      "title": "addNewLocationRequest",
      "example": {
        "ten_vi_tri": "Quận 1",
        "tinh_thanh": "Hồ Chí Minh",
        "hinh_anh": "https://airbnbnew.cybersoft.edu.vn/images/vt1.jpg"
      },
      "type": "object",
      "properties": {
        "ten_vi_tri": {
          "type": "string"
        },
        "tinh_thanh": {
          "type": "string"
        },
        "hinh_anh": {
          "type": "string"
        }
      },
      "required": [
        "ten_vi_tri",
        "tinh_thanh",
        "hinh_anh"
      ]
    },
...
  },
  "security": [],
...
}